requirements.txt
tqdm
numpy
pytorch
matplotlib
skimage